News-Channel-For-LLMS/scripts/requirements.txt
streamlit
streamlit-aggrid
sqlalchemy
psycopg2-binary